+++ dev/e2e_app/lib/main.dart
@@ -5,6 +5,7 @@ import 'package:e2e_app/location_screen.dart';
 import 'package:e2e_app/login_flow_screen.dart';
 import 'package:e2e_app/map_screen.dart';
 import 'package:e2e_app/notifications_screen.dart';
+import 'package:e2e_app/overflow_screen.dart';
 import 'package:e2e_app/overlay_screen.dart';
 import 'package:e2e_app/permissions_screen.dart';
 import 'package:e2e_app/scrolling_screen.dart';
@@ -99,14 +100,14 @@ class _ExampleHomePageState extends State<ExampleHomePage> {
         padding: EdgeInsets.all(8),
         key: const Key('listViewKey'),
         children: [
-          const Text('FIRST_KEY: ${const String.fromEnvironment('FIRST_KEY')}'),
+          const Text('FIRST_KEY: ${String.fromEnvironment('FIRST_KEY')}'),
           const Text(
-            'SECOND_KEY: ${const String.fromEnvironment('SECOND_KEY')}',
+            'SECOND_KEY: ${String.fromEnvironment('SECOND_KEY')}',
           ),
-          const Text('THIRD_KEY: ${const String.fromEnvironment('THIRD_KEY')}'),
-          const Text('FIFTH_KEY: ${const String.fromEnvironment('FIFTH_KEY')}'),
+          const Text('THIRD_KEY: ${String.fromEnvironment('THIRD_KEY')}'),
+          const Text('FIFTH_KEY: ${String.fromEnvironment('FIFTH_KEY')}'),
           const Text(
-            'BOOL_DEFINED: ${const String.fromEnvironment('BOOL_DEFINED')}',
+            'BOOL_DEFINED: ${String.fromEnvironment('BOOL_DEFINED')}',
           ),
           const Text(
             'You have pushed the button this many times:',
@@ -284,6 +285,14 @@ class _ExampleHomePageState extends State<ExampleHomePage> {
               ),
             ),
             child: const Text('Open login flow screen'),
+          ),
+          TextButton(
+            onPressed: () async => Navigator.of(context).push(
+              MaterialPageRoute<void>(
+                builder: (_) => const OverflowScreen(),
+              ),
+            ),
+            child: const Text('Open overflow screen'),
           ),
           Text('EXAMPLE_KEY: ${const String.fromEnvironment('EXAMPLE_KEY')}'),
         ],