+++ packages/patrol/lib/src/binding.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 import 'dart:convert';
 import 'dart:developer';
-import 'dart:io' as io;
+import 'dart:io';
 import 'dart:isolate';
 
 import 'package:flutter/foundation.dart';
@@ -15,6 +15,8 @@ import 'package:patrol/src/global_state.dart' as global_state;
 import 'constants.dart' as constants;
 
 const _success = 'success';
+
+bool get _isDevelopMode => constants.hotRestartEnabled;
 
 void _defaultPrintLogger(String message) {
   // TODO: Use a logger instead of print
@@ -44,30 +46,8 @@ class PatrolBinding extends LiveTestWidgetsFlutterBinding {
   /// You most likely don't want to call it yourself.
   PatrolBinding(NativeAutomatorConfig config)
       : _serviceExtensions = DevtoolsServiceExtensions(config) {
-    final oldTestExceptionReporter = reportTestException;
-
-    /// Wraps the default test exception reporter to report the test results to
-    /// the native side of Patrol.
-    reportTestException = (details, testDescription) {
-      final currentDartTest = _currentDartTest;
-      if (currentDartTest != null) {
-        assert(!constants.hotRestartEnabled);
-        // On iOS in release mode, diagnostics are compacted or truncated.
-        // We use the exceptionAsString() and stack to get the information
-        // about the exception. See [DiagnosticLevel].
-        final detailsAsString = (kReleaseMode && io.Platform.isIOS)
-            ? '${details.exceptionAsString()} \n ${details.stack}'
-            : details.toString();
-        _testResults[currentDartTest] = Failure(
-          testDescription,
-          detailsAsString,
-        );
-      }
-      oldTestExceptionReporter(details, testDescription);
-    };
-
     setUp(() {
-      if (constants.hotRestartEnabled) {
+      if (_isDevelopMode) {
         return;
       }
 
@@ -79,7 +59,7 @@ class PatrolBinding extends LiveTestWidgetsFlutterBinding {
     });
 
     tearDown(() async {
-      if (constants.hotRestartEnabled) {
+      if (_isDevelopMode) {
         // Sending results ends the test, which we don't want for Hot Restart
         return;
       }
@@ -231,11 +211,45 @@ class PatrolBinding extends LiveTestWidgetsFlutterBinding {
     Duration? timeout,
   }) async {
     await super.runTest(
-      testBody,
+      // If we're in develop mode, we cannot overwrite FlutterError.onError
+      // will get "A test overrode FlutterError.onError" exception.
+      _isDevelopMode
+          ? testBody
+          : () async => _wrapTestBodyWithExceptionGatherer(testBody),
       invariantTester,
       description: description,
     );
     _testResults[description] ??= _success;
+  }
+
+  /// Wraps the test body with a function that gathers exceptions and reports
+  /// them to the native side of Patrol.
+  Future<void> _wrapTestBodyWithExceptionGatherer(
+    Future<void> Function() testBody,
+  ) async {
+    final previousOnError = FlutterError.onError;
+    FlutterError.onError = (details) {
+      if (_currentDartTest case final testName?) {
+        final previousDetails = switch (_testResults[testName]) {
+          Failure(:final details?) => FlutterErrorDetails(exception: details),
+          _ => null,
+        };
+        final detailsAsString = (kReleaseMode && Platform.isIOS)
+            ? '${details.exceptionAsString()}\n${details.stack}'
+            : details.toString();
+
+        _testResults[testName] = Failure(
+          testName,
+          '$detailsAsString${previousDetails != null ? '\n$previousDetails' : ''}',
+        );
+
+        previousOnError?.call(details);
+      }
+    };
+
+    await testBody();
+
+    FlutterError.onError = previousOnError;
   }
 
   @override
@@ -250,12 +264,12 @@ class PatrolBinding extends LiveTestWidgetsFlutterBinding {
   @override
   Widget wrapWithDefaultView(Widget rootWidget) {
     assert(
-      (_currentDartTest != null) != (constants.hotRestartEnabled),
+      (_currentDartTest != null) != _isDevelopMode,
       '_currentDartTest can be null if and only if Hot Restart is enabled',
     );
 
     const testLabelEnabled = bool.fromEnvironment('PATROL_TEST_LABEL_ENABLED');
-    if (!testLabelEnabled || constants.hotRestartEnabled) {
+    if (!testLabelEnabled || _isDevelopMode) {
       return super.wrapWithDefaultView(RepaintBoundary(child: rootWidget));
     } else {
       return super.wrapWithDefaultView(